Moved some functionality from App to a landing page component
